+++ database/migrations/2026_02_08_044623_create_order_items_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('order_id')->constrained('orders')->cascadeOnDelete();
             $table->foreignId('product_id')->constrained('products')->cascadeOnDelete();
             $table->integer('quantity');
-            $table->float('price');
+            $table->decimal('price', 12, 2);
             $table->timestamps();
         });
     }
